Don't link dot if not present.
git-svn-id: https://ice.txcorp.com/svnrepos/code/bilder/trunk@2782 a9e16012-fd73-49ee-95cb-036de7809213

+++ packages/graphviz.sh
@@ -94,7 +94,9 @@ installGraphviz() {
     mkdir -p $CONTRIB_DIR/bin
     local instsfx=
     test $FORPYTHON_STATIC_BUILD != ser && instsfx="-${FORPYTHON_STATIC_BUILD}"
-    (cd $CONTRIB_DIR/bin; rm -f dot; ln -s ../graphviz${instsfx}/bin/dot .)
+    if test -f $CONTRIB_DIR/graphviz${instsfx}/bin/dot; then
+      (cd $CONTRIB_DIR/bin; rm -f dot; ln -s ../graphviz${instsfx}/bin/dot .)
+    fi
   fi
 }
